+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/TaskResolutionServices.kt
@@ -31,13 +31,8 @@ import org.eclipse.tractusx.bpdm.gate.entity.SyncTypeDb
 import org.eclipse.tractusx.bpdm.gate.entity.generic.BusinessPartnerDb
 import org.eclipse.tractusx.bpdm.gate.model.upsert.output.OutputUpsertData
 import org.eclipse.tractusx.bpdm.gate.repository.SharingStateRepository
-import org.eclipse.tractusx.bpdm.gate.repository.SyncRecordRepository
 import org.eclipse.tractusx.bpdm.gate.repository.generic.BusinessPartnerRepository
 import org.eclipse.tractusx.orchestrator.api.client.OrchestrationApiClient
-import org.eclipse.tractusx.orchestrator.api.model.ResultState
-import org.eclipse.tractusx.orchestrator.api.model.TaskClientStateDto
-import org.eclipse.tractusx.orchestrator.api.model.TaskResultStateSearchRequest
-import org.eclipse.tractusx.orchestrator.api.model.TaskStateRequest
 import org.eclipse.tractusx.orchestrator.api.model.*
 import org.springframework.data.domain.PageRequest
 import org.springframework.stereotype.Service
@@ -99,7 +94,6 @@ class TaskResolutionChunkService(
     private val businessPartnerService: BusinessPartnerService,
     private val orchestratorMappings: OrchestratorMappings,
     private val synchRecordService: SyncRecordService,
-    private val syncRecordRepository: SyncRecordRepository,
     private val goldenRecordUpdateService: GoldenRecordUpdateChunkService
 ) {
 
@@ -154,11 +148,7 @@ class TaskResolutionChunkService(
         resolveAsUpserts(successes)
         resolveAsErrors(errors)
 
-        events.content.lastOrNull()?.let { latestEvent ->
-            syncRecord.fromTime = latestEvent.timestamp
-            syncRecordRepository.save(syncRecord)
-        }
-
+        synchRecordService.updateRecord(syncRecord, events.content.lastOrNull()?.timestamp)
 
         logger.debug { "Resolved ${successes.size} tasks as successful, ${errors.size} as errors and ${unresolved.size} still unresolved" }